docker/deezer-dw/requirements.txt
paho-mqtt~=1.6.1
deezloader~=2022.3.5
deezloader~=2022.3.5
music-tag~=0.4.3